// src/types/Footer.types.ts
export type FooterType = {
    _id: any;
    title: any;
    logo: any;
    object: {
      companyName: any
      address: any;
      telephone: any;
      email: any;
      cvr: any;
    };
    social: {
      platform: any;
      url: any;
    }[];
    openingHours: {
      day: any;
      hours: any;
    }[];
  };
  